requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docx==0.2.4
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.58.0
fpdf==1.7.2
google==3.0.0
google_search_results==2.4.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
nltk==3.9.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
seaborn==0.13.2
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3


